Link Definitions: skip API version when resolving links
The BaseUrl currently contains the API version ('v2') as suffix.
The REST API returns links which contain the API version as prefix.
For example, /v2/groups/dtgd/28afe93ccf53e311b15d005056a10f14

This causes the version to be duplicated, as in the following:

21:39:53 connection.go:62: Sending request: GET /v2//v2/groups/dtgd/28afe93ccf53e311b15d005056a10f14 HTTP/1.1

The result is a 404 "Not Found" error.

Fixed by separating BaseUrl from API version.