Groups: remove Self()/Server()
The 'server' link is not always present within the Links array of a
group. It depends on whether the group contains servers. Hence the
call can fail with the error message

  "There is no link with rel server in model *groups.GetGroupRes"

Testing for this error in order to look for servers is brittle.
Hence remove this call, to be replaced by a different mechanism.

Also removed the Self() call, since this does not add any functio-
nality (identity operation). Continued documentation.